sec_recsys/csp/requirements.txt
appdirs==1.4.3
blinker==1.8.2
boto==2.49.0
boto3==1.35.82
botocore==1.35.82
bz2file==0.98
certifi==2018.10.15
chardet==3.0.4
charset-normalizer==3.4.0
click==8.1.7
cycler==0.10.0
debugpy==1.8.11
docutils==0.14
exceptiongroup==1.2.2
fixedpoint==1.0.1
Flask==3.0.3
Flask-JSON==0.4.0
funcy==2.0
future==1.0.0
idna==2.7
importlib_metadata==8.5.0
iniconfig==2.0.0
itsdangerous==2.2.0
Jinja2==3.1.4
jmespath==0.9.3
joblib==1.4.2
jsonify==0.5
kiwisolver==1.0.1
MarkupSafe==2.1.5
mpmath==1.3.0
nltk==3.9.1
numexpr==2.8.6
numpy==1.22.2
packaging==16.8
pandas==1.4.1
pillow==10.4.0
pluggy==1.5.0
psycopg2-binary==2.9.10
pycryptodomex==3.21.0
pyparsing==2.3.0
pytest==8.3.4
python-dateutil==2.8.2
pytz==2021.3
regex==2024.11.6
requests==2.32.3
s3transfer==0.10.4
six==1.11.0
smart-open==7.0.5
sqlparse==0.5.3
sympy==1.13.3
tenseal==0.3.15
threadpoolctl==3.5.0
tomli==2.2.1
tqdm==4.67.1
urllib3==1.26.20
Werkzeug==3.0.6
wrapt==1.17.0
zipp==3.20.2
